\name{ccTutorial-package}
\alias{ccTutorial-package}
\alias{ccTutorial}
\docType{package}
\title{
What the package does (short line)
}
\description{
 Data package for a Bioconductor ChIP-chip tutorial
}
\details{
\tabular{ll}{
Package: \tab ccTutorial\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-04-29\cr
License: \tab Artistic License 2.0\cr
}
}
\author{
Joern Toedling, Wolfgang Huber

Maintainer: Joern Toedling <toedling@ebi.ac.uk>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
