\name{chersX}
\alias{chersX}
\docType{data}
\title{List of found ChIP-enriched regions}
\description{
  This is the list of found ChIP-enriched regions.
  The package contains all the data to recreate the list
  and this list is merely given for convenience sake.
}
\usage{data(chersX)}
\format{
  A \code{cherList}, a list of \code{cher} objects.
}
%\details{}
%\source{}
%\references{}
\seealso{\code{\link[Ringo:cherClass]{cher-class}}}
\examples{
data(chersX)
head(chersX)
}
\keyword{datasets}
